2/7 more camera work -ECS
