+++ Sensor/Python/camera stuff.json
@@ -1,0 +1,135 @@
+{
+  "baseline": "-50.0087",
+  "intrinsic_left.x.x": "0.499694",
+  "intrinsic_left.x.y": "0.799109",
+  "intrinsic_left.x.z": "0.49929",
+  "intrinsic_left.y.x": "0.508718",
+  "intrinsic_left.y.y": "-0.0557761",
+  "intrinsic_left.y.z": "0.0628533",
+  "intrinsic_left.z.x": "-0.000276484",
+  "intrinsic_left.z.y": "-0.00109871",
+  "intrinsic_left.z.z": "-0.020306",
+  "intrinsic_right.x.x": "0.497527",
+  "intrinsic_right.x.y": "0.795203",
+  "intrinsic_right.x.z": "0.504607",
+  "intrinsic_right.y.x": "0.506111",
+  "intrinsic_right.y.y": "-0.0572176",
+  "intrinsic_right.y.z": "0.069096",
+  "intrinsic_right.z.x": "0.000487369",
+  "intrinsic_right.z.y": "-0.000195329",
+  "intrinsic_right.z.z": "-0.0225608",
+  "rectified.0.fx": "964.4",
+  "rectified.0.fy": "964.4",
+  "rectified.0.height": "1080",
+  "rectified.0.ppx": "970.997",
+  "rectified.0.ppy": "548.426",
+  "rectified.0.width": "1920",
+  "rectified.1.fx": "642.933",
+  "rectified.1.fy": "642.933",
+  "rectified.1.height": "720",
+  "rectified.1.ppx": "647.332",
+  "rectified.1.ppy": "365.617",
+  "rectified.1.width": "1280",
+  "rectified.10.fx": "578.64",
+  "rectified.10.fy": "578.64",
+  "rectified.10.height": "0",
+  "rectified.10.ppx": "366.598",
+  "rectified.10.ppy": "365.056",
+  "rectified.10.width": "0",
+  "rectified.11.fx": "462.912",
+  "rectified.11.fy": "462.912",
+  "rectified.11.height": "0",
+  "rectified.11.ppx": "293.279",
+  "rectified.11.ppy": "292.044",
+  "rectified.11.width": "0",
+  "rectified.12.fx": "0",
+  "rectified.12.fy": "0",
+  "rectified.12.height": "400",
+  "rectified.12.ppx": "0",
+  "rectified.12.ppy": "0",
+  "rectified.12.width": "640",
+  "rectified.13.fx": "0",
+  "rectified.13.fy": "0",
+  "rectified.13.height": "576",
+  "rectified.13.ppx": "0",
+  "rectified.13.ppy": "0",
+  "rectified.13.width": "576",
+  "rectified.14.fx": "0",
+  "rectified.14.fy": "0",
+  "rectified.14.height": "720",
+  "rectified.14.ppx": "0",
+  "rectified.14.ppy": "0",
+  "rectified.14.width": "720",
+  "rectified.15.fx": "0",
+  "rectified.15.fy": "0",
+  "rectified.15.height": "1152",
+  "rectified.15.ppx": "0",
+  "rectified.15.ppy": "0",
+  "rectified.15.width": "1152",
+  "rectified.2.fx": "385.76",
+  "rectified.2.fy": "385.76",
+  "rectified.2.height": "480",
+  "rectified.2.ppx": "324.399",
+  "rectified.2.ppy": "243.37",
+  "rectified.2.width": "640",
+  "rectified.3.fx": "425.943",
+  "rectified.3.fy": "425.943",
+  "rectified.3.height": "480",
+  "rectified.3.ppx": "428.857",
+  "rectified.3.ppy": "243.721",
+  "rectified.3.width": "848",
+  "rectified.4.fx": "321.467",
+  "rectified.4.fy": "321.467",
+  "rectified.4.height": "360",
+  "rectified.4.ppx": "323.666",
+  "rectified.4.ppy": "182.809",
+  "rectified.4.width": "640",
+  "rectified.5.fx": "212.972",
+  "rectified.5.fy": "212.972",
+  "rectified.5.height": "240",
+  "rectified.5.ppx": "214.429",
+  "rectified.5.ppy": "121.861",
+  "rectified.5.width": "424",
+  "rectified.6.fx": "192.88",
+  "rectified.6.fy": "192.88",
+  "rectified.6.height": "240",
+  "rectified.6.ppx": "162.199",
+  "rectified.6.ppy": "121.685",
+  "rectified.6.width": "320",
+  "rectified.7.fx": "241.1",
+  "rectified.7.fy": "241.1",
+  "rectified.7.height": "270",
+  "rectified.7.ppx": "242.749",
+  "rectified.7.ppy": "137.107",
+  "rectified.7.width": "480",
+  "rectified.8.fx": "642.933",
+  "rectified.8.fy": "642.933",
+  "rectified.8.height": "800",
+  "rectified.8.ppx": "647.332",
+  "rectified.8.ppy": "405.617",
+  "rectified.8.width": "1280",
+  "rectified.9.fx": "482.2",
+  "rectified.9.fy": "482.2",
+  "rectified.9.height": "540",
+  "rectified.9.ppx": "485.499",
+  "rectified.9.ppy": "274.213",
+  "rectified.9.width": "960",
+  "world2left_rot.x.x": "0.999964",
+  "world2left_rot.x.y": "-0.00122291",
+  "world2left_rot.x.z": "0.00837479",
+  "world2left_rot.y.x": "0.00120143",
+  "world2left_rot.y.y": "0.999996",
+  "world2left_rot.y.z": "0.00256966",
+  "world2left_rot.z.x": "-0.00837789",
+  "world2left_rot.z.y": "-0.00255951",
+  "world2left_rot.z.z": "0.999962",
+  "world2right_rot.x.x": "0.999997",
+  "world2right_rot.x.y": "-0.000260638",
+  "world2right_rot.x.z": "0.00237314",
+  "world2right_rot.y.x": "0.000266723",
+  "world2right_rot.y.y": "0.999997",
+  "world2right_rot.y.z": "-0.00256429",
+  "world2right_rot.z.x": "-0.00237247",
+  "world2right_rot.z.y": "0.00256492",
+  "world2right_rot.z.z": "0.999994"
+}
