--- Sensor/Python/realsense_depth.py
@@ -1,51 +1,0 @@
-import pyrealsense2 as rs
-import numpy as np
-
-class DepthCamera:
-    def __init__(self):
-        # Configure depth and color streams
-        self.pipeline = rs.pipeline()
-        config = rs.config()
-
-        # Get device product line for setting a supporting resolution
-        pipeline_wrapper = rs.pipeline_wrapper(self.pipeline)
-        pipeline_profile = config.resolve(pipeline_wrapper)
-        device = pipeline_profile.get_device()
-        device_product_line = str(device.get_info(rs.camera_info.product_line))
-
-        config.enable_stream(rs.stream.depth, 640, 480, rs.format.z16, 30)
-        config.enable_stream(rs.stream.color, 640, 480, rs.format.bgr8, 30)
-
-        # Start streaming
-        self.pipeline.start(config)
-
-    def get_frame(self):
-        frames = self.pipeline.wait_for_frames()
-        depth_frame = frames.get_depth_frame()
-        color_frame = frames.get_color_frame()
-                
-        
-        depth_image = np.asanyarray(depth_frame.get_data())
-        color_image = np.asanyarray(color_frame.get_data())
-        if not depth_frame or not color_frame:
-            return False, None, None
-        return True, depth_image, color_image
-
-    def convert_depth_to_phys_coord_using_realsense(x, y, depth, cameraInfo):  
-        _intrinsics = rs.intrinsics()
-        _intrinsics.width = cameraInfo.width
-        _intrinsics.height = cameraInfo.height
-        _intrinsics.ppx = cameraInfo.K[2]
-        _intrinsics.ppy = cameraInfo.K[5]
-        _intrinsics.fx = cameraInfo.K[0]
-        _intrinsics.fy = cameraInfo.K[4]
-        #_intrinsics.model = cameraInfo.distortion_model
-        _intrinsics.model  = rs.distortion.none  
-        _intrinsics.coeffs = [i for i in cameraInfo.D]  
-    
-        result = rs.rs2_deproject_pixel_to_point(_intrinsics, [x, y], depth)  #result[0]: right, result[1]: down, result[2]: forward
-        return result[2], -result[0], -result[1]
-        
-
-    def release(self):
-        self.pipeline.stop()
